+++ packages/verifiable-credentials/test/utils.ts
@@ -1,6 +1,6 @@
-
 import { AutoAccount } from '@verida/account-node';
-import { Client, Context, Network, Utils } from '@verida/client-ts';
+import { Client, Context, Network } from '@verida/client-ts';
+import { explodeVeridaUri } from '@verida/helpers';
 import { ClientConfig, EnvironmentType } from '@verida/types'
 import CONFIG from './config';
 
@@ -20,8 +20,6 @@ export async function connectAccount(privateKey: string, contextName: string, en
             environment: EnvironmentType.TESTNET,
         }
     )
-
-    const did = await account.did()
 
     const context = await Network.connect({
         context: {
@@ -43,16 +41,15 @@ export async function connectAccount(privateKey: string, contextName: string, en
  * @returns an application {context} of the connected account
  */
 export async function getClientContext(uri: string, environment: EnvironmentType): Promise<Context> {
-
     const clientConfig: ClientConfig = {
         environment: environment
     }
 
-    const url = Utils.explodeVeridaUri(uri)
+    const uriParts = explodeVeridaUri(uri)
 
     const context = await new Client(clientConfig).openExternalContext(
-        url.contextName,
-        url.did
+        uriParts.contextName,
+        uriParts.did
     );
 
     return <Context> context